+++ blocks/email-composer/email-composer.css
@@ -65,11 +65,23 @@
     flex: 1;
 }
 
-.email-composer.block .preview iframe {
+.email-composer.block .preview .iframe {
     width: 100%;
     height: 100%;
-    border: 1px solid var(--gray);
+    border: 1px solid var(--light-gray);
     transition: all var(--animation-duration);
+}
+
+.email-composer.block .iframe.is-loading {
+    display: none;
+}
+
+.email-composer.block .iframe + .skeleton {
+    border-radius: 0;
+}
+
+.email-composer.block .iframe:not(.is-loading) + .skeleton {
+    display: none;
 }
 
 .email-composer input:is([type="text"], [type="email"]) {
@@ -151,7 +163,7 @@
         flex-direction: column;
     }
 
-    .email-composer.block .preview iframe {
+    .email-composer.block .preview .iframe {
         height: 100vh;
     }
