+++ blocks/theme-editor/theme-editor.css
@@ -94,7 +94,7 @@
     width: 100%;
     height: 100%;
     transition: all var(--animation-duration);
-    border: 1px solid var(--gray);
+    border: 1px solid var(--light-gray);
     max-width: 1440px;
 }
